refactor: 내부 controller와 외부 controller 구분
